# docs/movie.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "movie.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MovieService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/movie": {
      "get": {
        "operationId": "MovieService_SearchMovie",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbFindAllResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "pagination",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "searchword",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "MovieService"
        ]
      }
    },
    "/v1/movie/{id}": {
      "get": {
        "operationId": "MovieService_SearchMovieByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbMovie"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "MovieService"
        ]
      }
    }
  },
  "definitions": {
    "pbFindAllResponse": {
      "type": "object",
      "properties": {
        "Results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbMovieList"
          }
        },
        "TotalResult": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbMovie": {
      "type": "object",
      "properties": {
        "Title": {
          "type": "string"
        },
        "Year": {
          "type": "string"
        },
        "Rated": {
          "type": "string"
        },
        "Released": {
          "type": "string"
        },
        "Runtime": {
          "type": "string"
        },
        "Genre": {
          "type": "string"
        },
        "Director": {
          "type": "string"
        },
        "Writer": {
          "type": "string"
        },
        "Actors": {
          "type": "string"
        },
        "Plot": {
          "type": "string"
        },
        "Language": {
          "type": "string"
        },
        "Country": {
          "type": "string"
        },
        "Awards": {
          "type": "string"
        },
        "Poster": {
          "type": "string"
        },
        "Ratings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/pbRating"
          }
        },
        "Metascore": {
          "type": "string"
        },
        "imdbRating": {
          "type": "string"
        },
        "imdbVotes": {
          "type": "string"
        },
        "imdbID": {
          "type": "string"
        },
        "Type": {
          "type": "string"
        },
        "totalSeasons": {
          "type": "string"
        },
        "Response": {
          "type": "string"
        }
      }
    },
    "pbMovieList": {
      "type": "object",
      "properties": {
        "Title": {
          "type": "string"
        },
        "Year": {
          "type": "string"
        },
        "imdbID": {
          "type": "string"
        },
        "Type": {
          "type": "string"
        },
        "Poster": {
          "type": "string"
        }
      }
    },
    "pbRating": {
      "type": "object",
      "properties": {
        "Source": {
          "type": "string"
        },
        "Value": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}